+++ libs/gesha-core/src/languages/mod.rs
@@ -1,0 +1,1 @@
+pub mod rust;
