+++ libs/gesha-core/src/v3_0/to_rust/to_struct.rs
@@ -1,4 +1,4 @@
-use crate::rust_types::{Definition, StructField};
+use crate::languages::rust::{Definition, StructField};
 use openapi_types::v3_0::{SchemaCase, SchemaFieldName, SchemaObject, SchemaProperties};
 
 pub(super) fn to_struct(name: SchemaFieldName, object: SchemaObject) -> crate::Result<Definition> {
